+++ .gitignore
@@ -36,3 +36,6 @@ Carthage/Build
 
 # Other
 ._*
+
+# Temp
+02_knight_and_princess/tevfik_aksoy/temp/
